+++ src/main/java/de/markusbordihn/easynpc/client/screen/configuration/skin/DefaultSkinConfigurationScreen.java
@@ -22,9 +22,6 @@ package de.markusbordihn.easynpc.client.screen.configuration.skin;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-
 import com.mojang.blaze3d.systems.RenderSystem;
 import com.mojang.blaze3d.vertex.PoseStack;
 
@@ -49,8 +46,6 @@ import de.markusbordihn.easynpc.utils.TextUtils;
 @OnlyIn(Dist.CLIENT)
 public class DefaultSkinConfigurationScreen
     extends SkinConfigurationScreen<DefaultSkinConfigurationMenu> {
-
-  protected static final Logger log = LogManager.getLogger(Constants.LOG_NAME);
 
   // Internal
   private Button skinPreviousButton = null;
@@ -134,7 +129,7 @@ public class DefaultSkinConfigurationScreen
         });
 
     // Render active skin in different style.
-    if (this.entity.getSkinType() == SkinType.CUSTOM && this.entity.getVariant().equals(variant)
+    if (this.entity.getSkinType() == SkinType.DEFAULT && this.entity.getVariant().equals(variant)
         && (profession == null || this.entity.getProfession().equals(profession))) {
       poseStack.pushPose();
       RenderSystem.setShader(GameRenderer::getPositionTexShader);
@@ -173,7 +168,6 @@ public class DefaultSkinConfigurationScreen
     super.init();
 
     // Default button stats
-    this.closeButton.visible = false;
     this.customSkinButton.active = true;
     this.defaultSkinButton.active = false;
     this.playerSkinButton.active = true;