+++ src/main/java/de/markusbordihn/easynpc/network/message/MessageActionChange.java
@@ -25,43 +25,50 @@ import java.util.function.Supplier;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
+import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.level.ServerPlayer;
 
 import net.minecraftforge.network.NetworkEvent;
 
 import de.markusbordihn.easynpc.Constants;
+import de.markusbordihn.easynpc.action.ActionType;
 import de.markusbordihn.easynpc.entity.EasyNPCEntity;
-import de.markusbordihn.easynpc.entity.EasyNPCEntityMenu;
 import de.markusbordihn.easynpc.entity.EntityManager;
 
-public class MessageOpenDialog {
+public class MessageActionChange {
 
   protected static final Logger log = LogManager.getLogger(Constants.LOG_NAME);
 
+  protected final String action;
+  protected final String actionType;
   protected final UUID uuid;
-  protected final String dialogName;
 
-  public MessageOpenDialog(UUID uuid, String dialogName) {
+  public MessageActionChange(UUID uuid, String actionType, String action) {
     this.uuid = uuid;
-    this.dialogName = dialogName;
+    this.actionType = actionType;
+    this.action = action;
   }
 
-  public String getDialogName() {
-    return this.dialogName;
+  public String getActionType() {
+    return this.actionType;
+  }
+
+  public String getAction() {
+    return this.action;
   }
 
   public UUID getUUID() {
     return this.uuid;
   }
 
-  public static void handle(MessageOpenDialog message,
+  public static void handle(MessageActionChange message,
       Supplier<NetworkEvent.Context> contextSupplier) {
     NetworkEvent.Context context = contextSupplier.get();
     context.enqueueWork(() -> handlePacket(message, context));
     context.setPacketHandled(true);
   }
 
-  public static void handlePacket(MessageOpenDialog message, NetworkEvent.Context context) {
+  public static void handlePacket(MessageActionChange message, NetworkEvent.Context context) {
     ServerPlayer serverPlayer = context.getSender();
     if (serverPlayer == null) {
       log.error("Unable to get server player for message {} from {}", message, context);
@@ -75,10 +82,10 @@ public class MessageOpenDialog {
       return;
     }
 
-    // Validate dialog name.
-    String dialogName = message.getDialogName();
-    if (dialogName == null || dialogName.isEmpty()) {
-      log.error("Invalid dialog name {} for {} from {}", dialogName, message, serverPlayer);
+    // Validate action type.
+    ActionType actionType = ActionType.get(message.getActionType());
+    if (actionType == null || actionType == ActionType.NONE) {
+      log.error("Invalid action type {} for {} from {}", actionType, message, serverPlayer);
       return;
     }
 
@@ -89,29 +96,19 @@ public class MessageOpenDialog {
       return;
     }
 
-    // Perform action.
-    switch (dialogName) {
-      case "BasicActionConfiguration":
-        EasyNPCEntityMenu.openBasicActionConfigurationMenu(serverPlayer, easyNPCEntity);
-      break;
-      case "BasicDialogConfiguration":
-        EasyNPCEntityMenu.openBasicDialogConfigurationMenu(serverPlayer, easyNPCEntity);
-        break;
-      case "YesNoDialogConfiguration":
-        EasyNPCEntityMenu.openYesNoDialogConfigurationMenu(serverPlayer, easyNPCEntity);
-        break;
-      case "CustomSkinConfiguration":
-        EasyNPCEntityMenu.openCustomSkinConfigurationMenu(serverPlayer, easyNPCEntity);
-        break;
-      case "DefaultSkinConfiguration":
-        EasyNPCEntityMenu.openDefaultSkinConfigurationMenu(serverPlayer, easyNPCEntity);
-        break;
-      case "PlayerSkinConfiguration":
-        EasyNPCEntityMenu.openPlayerSkinConfigurationMenu(serverPlayer, easyNPCEntity);
-        break;
-      default:
-        log.debug("Unknown dialog {} for {} from {}", dialogName, easyNPCEntity, serverPlayer);
+    // Get Permission level for corresponding action.
+    int permissionLevel = 0;
+    MinecraftServer minecraftServer = serverPlayer.getServer();
+    if (minecraftServer != null) {
+      permissionLevel = minecraftServer.getProfilePermissions(serverPlayer.getGameProfile());
+      easyNPCEntity.setActionPermissionLevel(permissionLevel);
     }
+
+    // Perform action.
+    String action = message.getAction();
+    log.debug("Set action {}:{} for {} from {} with permission level {}.", actionType, action,
+        easyNPCEntity, serverPlayer, permissionLevel);
+    easyNPCEntity.setAction(actionType, action);
   }
 
 }